Feature/optional packages (#6)
* Add missing branch trigger

* Test minimal base packages

* Split readers into own files and strip down default package requirments

* tidy up warning

* dont warn
