+++ .github/workflows/code-quality.yml
@@ -10,6 +10,7 @@ on:
   push:
     branches:
       - main
+      - develop  
       - stable**
 
 env:
