+++ oasis_data_manager/df_reader/exceptions.py
@@ -1,2 +1,6 @@
 class InvalidSQLException(Exception):
     pass
+
+
+class MissingOptionalDependency(Exception):
+    pass
