+++ optional-package.in
@@ -1,0 +1,9 @@
+adlfs
+boto3
+dask
+dask-geopandas
+dask-sql
+distributed
+geopandas
+pyogrio
+s3fs>=2023.9.0
